Remove MediatR from all skills
