+++ Guppi.Application/DependencyInjection.cs
@@ -1,22 +1,18 @@
-using System.Reflection;
 using Guppi.Application.Services;
-using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Guppi.Application;
 
 public static class DependencyInjection
 {
-    public static IServiceCollection AddApplication(this IServiceCollection services)
-    {
-        services
-            .AddMediatR(Assembly.GetExecutingAssembly())
+    public static IServiceCollection AddApplication(this IServiceCollection services) => services
             .AddTransient<IAdventOfCodeService, AdventOfCodeService>()
             .AddTransient<IAsciiService, AsciiService>()
             .AddTransient<ICalendarService, CalendarService>()
             .AddTransient<IGitService, GitService>()
-            .AddTransient<IHueLightService, HueLightService>();
-
-        return services;
-    }
+            .AddTransient<IHueLightService, HueLightService>()
+            .AddTransient<INoteService, NoteService>()
+            .AddTransient<IStravaService, StravaService>()
+            .AddTransient<IUtilitiesService, UtilitiesService>()
+            .AddTransient<IWeatherService, WeatherService>();
 }
