+++ Guppi.Application/Services/IUtilitiesService.cs
@@ -1,0 +1,8 @@
+using System.Threading.Tasks;
+
+namespace Guppi.Application.Services;
+
+public interface IUtilitiesService
+{
+    Task RestartExplorer();
+}
