+++ Guppi.Console/Skills/NotesSkill.cs
@@ -1,20 +1,18 @@
-using System;
 using System.Collections.Generic;
 using System.CommandLine;
 using System.CommandLine.NamingConventionBinder;
 using System.Threading.Tasks;
-using Guppi.Application.Commands.Notes;
-using MediatR;
+using Guppi.Application.Services;
 
 namespace Guppi.Console.Skills
 {
     public class NotesSkill : ISkill
     {
-        private readonly IMediator _mediator;
+        private readonly INoteService _service;
 
-        public NotesSkill(IMediator mediator)
+        public NotesSkill(INoteService service)
         {
-            _mediator = mediator;
+            _service = service;
         }
 
         public IEnumerable<Command> GetCommands()
@@ -25,27 +23,27 @@ namespace Guppi.Console.Skills
                 //new Option<string>( new string[]{"--folder", "-f"}, () => string.Empty, "Sub-folder for the note"),
                 new Option<string>( new string[]{"--vault", "-v"}, () => string.Empty, "Use a vault other than the default")
             };
-            add.Handler = CommandHandler.Create(async (string title, /* string folder, */ string vault) => await Add(title, null, vault));
+            add.Handler = CommandHandler.Create((string title, /* string folder, */ string vault) => Add(title, null, vault));
 
             //var daily = new Command("daily", "Adds a new daily note")
             //{
             //    new Option<string>( new string[]{"--vault", "-v"}, () => string.Empty, "Use a vault other than the default")
             //};
-            //daily.Handler = CommandHandler.Create(async (string vault) => await Daily(vault));
+            //daily.Handler = CommandHandler.Create(async (string vault) => Daily(vault));
 
             var view = new Command("view", "Opens Obsidian")
             {
                 new Option<string>( new string[]{"--vault", "-v"}, () => string.Empty, "Use a vault other than the default")
             };
             view.AddAlias("open");
-            view.Handler = CommandHandler.Create(async (string vault) => await View(vault));
+            view.Handler = CommandHandler.Create((string vault) => View(vault));
 
             var code = new Command("code", "Opens the notes folder in VS Code");
-            code.Handler = CommandHandler.Create(async () => await Code());
+            code.Handler = CommandHandler.Create(() => Code());
 
             var config = new Command("configure", "Configure the notes skill.");
             config.AddAlias("config");
-            config.Handler = CommandHandler.Create(async () => await Configure());
+            config.Handler = CommandHandler.Create(() => Configure());
 
             var notes = new Command("notes", "Works with Notes in Obsidian")
             {
@@ -58,8 +56,8 @@ namespace Guppi.Console.Skills
             return new[] { notes };
         }
 
-        private async Task Add(string title, string folder, string vault) =>
-            await _mediator.Send(new AddFileCommand(title, folder, vault));
+        private void Add(string title, string folder, string vault) =>
+            _service.AddFile(vault, title);
 
         private Task Daily(string vault)
         {
@@ -67,13 +65,13 @@ namespace Guppi.Console.Skills
             return Task.CompletedTask;
         }
 
-        private async Task View(string vault) =>
-            await _mediator.Send(new OpenObsidianCommand(vault, null));
+        private void View(string vault) =>
+            _service.OpenObsidian(vault, null);
 
-        private async Task Code() => 
-            await _mediator.Send(new OpenCodeCommand());
+        private void Code() => 
+            _service.OpenVsCode();
 
-        private async Task Configure() => 
-            await _mediator.Send(new ConfigureNotesCommand());
+        private void Configure() =>
+            _service.Configure();
     }
 }
